refactor: destructurize
