+++ src/widgets/Catalog/filters/field/types.ts
@@ -1,0 +1,4 @@
+export type TFiltersField = {
+  id: string;
+  name: string;
+};
