+++ src/widgets/Catalog/filters/styles.ts
@@ -1,0 +1,8 @@
+import { styled } from 'styled-components';
+
+export const WrapFilters = styled.div`
+  margin-right: 1.5rem;
+  li {
+    margin: 0.5rem 0;
+  }
+`;
